backend/requirements.txt
# Django Backend
django
djangorestframework
django-cors-headers
python-decouple
Pillow
djangorestframework-simplejwt

# Machine Learning & Data Science
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow
torch
torchvision
jupyter
notebook
opencv-python

# HTTP & API
requests
httpx
groq

# Development & Testing
pytest
pytest-django
black
flake8
python-dotenv

# Database (additional SQLite tools)
sqlite-utils

# Utilities
python-dateutil
pytz